+++ src/tests/wallet_test.ts
@@ -279,13 +279,6 @@ describe("Wallet Class", () => {
         expect(op).toBeInstanceOf(StakingOperation);
       });
 
-      it("should throw an error when the wallet does not have a default address", async () => {
-        const newWallet = Wallet.init(walletModel);
-        await expect(
-          async () => await newWallet.createStake(0.001, Coinbase.assets.Eth),
-        ).rejects.toThrow(Error);
-      });
-
       it("should throw an error when wait is called on wallet address based staking operation", async () => {
         const wallet = await Wallet.create({ networkId: Coinbase.networks.EthereumHolesky });
         const op = await wallet.createStake(0.001, Coinbase.assets.Eth);
@@ -318,13 +311,6 @@ describe("Wallet Class", () => {
 
         expect(op).toBeInstanceOf(StakingOperation);
       });
-
-      it("should throw an error when the wallet does not have a default address", async () => {
-        const newWallet = Wallet.init(walletModel);
-        await expect(
-          async () => await newWallet.createUnstake(0.001, Coinbase.assets.Eth),
-        ).rejects.toThrow(Error);
-      });
     });
 
     describe(".createClaimStake", () => {
@@ -345,23 +331,9 @@ describe("Wallet Class", () => {
 
         expect(op).toBeInstanceOf(StakingOperation);
       });
-
-      it("should throw an error when the wallet does not have a default address", async () => {
-        const newWallet = Wallet.init(walletModel);
-        await expect(
-          async () => await newWallet.createClaimStake(0.001, Coinbase.assets.Eth),
-        ).rejects.toThrow(Error);
-      });
     });
 
     describe(".stakeableBalance", () => {
-      it("should throw an error when the wallet does not have a default address", async () => {
-        const newWallet = Wallet.init(walletModel);
-        await expect(
-          async () => await newWallet.stakeableBalance(Coinbase.assets.Eth),
-        ).rejects.toThrow(Error);
-      });
-
       it("should return the stakeable balance successfully with default params", async () => {
         //const wallet = await Wallet.create({ networkId: Coinbase.networks.EthereumHolesky });
         Coinbase.apiClients.stake!.getStakingContext = mockReturnValue(STAKING_CONTEXT_MODEL);
@@ -371,13 +343,6 @@ describe("Wallet Class", () => {
     });
 
     describe(".unstakeableBalance", () => {
-      it("should throw an error when the wallet does not have a default address", async () => {
-        const newWallet = Wallet.init(walletModel);
-        await expect(
-          async () => await newWallet.unstakeableBalance(Coinbase.assets.Eth),
-        ).rejects.toThrow(Error);
-      });
-
       it("should return the unstakeableBalance balance successfully with default params", async () => {
         const wallet = await Wallet.create({ networkId: Coinbase.networks.EthereumHolesky });
         Coinbase.apiClients.stake!.getStakingContext = mockReturnValue(STAKING_CONTEXT_MODEL);
@@ -387,13 +352,6 @@ describe("Wallet Class", () => {
     });
 
     describe(".claimableBalance", () => {
-      it("should throw an error when the wallet does not have a default address", async () => {
-        const newWallet = Wallet.init(walletModel);
-        await expect(
-          async () => await newWallet.claimableBalance(Coinbase.assets.Eth),
-        ).rejects.toThrow(Error);
-      });
-
       it("should return the claimableBalance balance successfully with default params", async () => {
         const wallet = await Wallet.create({ networkId: Coinbase.networks.EthereumHolesky });
         Coinbase.apiClients.stake!.getStakingContext = mockReturnValue(STAKING_CONTEXT_MODEL);
@@ -403,13 +361,6 @@ describe("Wallet Class", () => {
     });
 
     describe(".stakingRewards", () => {
-      it("should throw an error when the wallet does not have a default address", async () => {
-        const newWallet = Wallet.init(walletModel);
-        await expect(
-          async () => await newWallet.stakingRewards(Coinbase.assets.Eth),
-        ).rejects.toThrow(Error);
-      });
-
       it("should successfully return staking rewards", async () => {
         const wallet = await Wallet.create({ networkId: Coinbase.networks.EthereumHolesky });
         Coinbase.apiClients.stake!.fetchStakingRewards = mockReturnValue(STAKING_REWARD_RESPONSE);
@@ -420,13 +371,6 @@ describe("Wallet Class", () => {
     });
 
     describe(".historicalStakingBalances", () => {
-      it("should throw an error when the wallet does not have a default address", async () => {
-        const newWallet = Wallet.init(walletModel);
-        await expect(
-          async () => await newWallet.historicalStakingBalances(Coinbase.assets.Eth),
-        ).rejects.toThrow(Error);
-      });
-
       it("should successfully return historical staking balances", async () => {
         const wallet = await Wallet.create({ networkId: Coinbase.networks.EthereumHolesky });
         Coinbase.apiClients.stake!.fetchHistoricalStakingBalances = mockReturnValue(
@@ -484,16 +428,6 @@ describe("Wallet Class", () => {
       Coinbase.apiClients.externalAddress!.listAddressHistoricalBalance = mockReturnValue(
         mockHistoricalBalanceResponse,
       );
-    });
-
-    it("should throw an error when the wallet does not have a default address", async () => {
-      const newWallet = Wallet.init(walletModel);
-      await expect(
-        async () =>
-          await newWallet.listHistoricalBalances({
-            assetId: Coinbase.assets.Usdc,
-          }),
-      ).rejects.toThrow(Error);
     });
 
     it("should successfully return historical balances", async () => {
@@ -644,18 +578,6 @@ describe("Wallet Class", () => {
       expect(contractInvocation).toBeInstanceOf(ContractInvocation);
       expect(contractInvocation).toEqual(expectedInvocation);
     });
-
-    describe("when the wallet does not have a default address", () => {
-      let invalidWallet;
-
-      beforeEach(() => (invalidWallet = Wallet.init(walletModel)));
-
-      it("should throw an Error", async () => {
-        await expect(async () => await invalidWallet.invokeContract(options)).rejects.toThrow(
-          Error,
-        );
-      });
-    });
   });
 
   describe("#createPayloadSignature", () => {
@@ -708,16 +630,6 @@ describe("Wallet Class", () => {
         },
       );
       expect(Coinbase.apiClients.address!.createPayloadSignature).toHaveBeenCalledTimes(1);
-    });
-
-    it("should throw an Error when the wallet does not have a default address", async () => {
-      const invalidWallet = Wallet.init(walletModel);
-
-      expect(async () => {
-        await invalidWallet.createPayloadSignature(unsignedPayload);
-      }).rejects.toThrow(Error);
-
-      expect(Coinbase.apiClients.address!.createPayloadSignature).not.toHaveBeenCalled();
     });
   });
 
@@ -782,16 +694,6 @@ describe("Wallet Class", () => {
     describe("#getDefaultAddress", () => {
       it("should return the correct default address", async () => {
         expect((await wallet.getDefaultAddress()).getId()).toBe(walletModel.default_address!.address_id);
-      });
-
-      it("should throw an Error if the wallet does not have a default address", async () => {
-        const walletModelNoDefaultAddress = {
-          id: walletId,
-          network_id: Coinbase.networks.BaseSepolia,
-          feature_set: {} as FeatureSet,
-        };
-        const newWallet = Wallet.init(walletModelNoDefaultAddress);
-        await expect(newWallet.getDefaultAddress()).rejects.toThrow("WalletModel default address not set");
       });
     });
 
@@ -920,15 +822,6 @@ describe("Wallet Class", () => {
           },
         };
       });
-    });
-
-    it("should throw an error when the wallet does not have a default address", async () => {
-      const wallet = Wallet.init({
-        id: walletId,
-        network_id: Coinbase.networks.BaseSepolia,
-        feature_set: {} as FeatureSet,
-      });
-      await expect(async () => await wallet.faucet()).rejects.toThrow(Error);
     });
 
     it("should return a Wallet instance", async () => {
@@ -1348,14 +1241,6 @@ describe("Wallet Class", () => {
         status: TransactionStatusEnum.Pending,
       },
     } as TradeModel);
-
-    it("should throw an error when the wallet does not have a default address", async () => {
-      const newWallet = Wallet.init(walletModel);
-      await expect(
-        async () =>
-          await newWallet.createTrade({ amount: 0.01, fromAssetId: "eth", toAssetId: "usdc" }),
-      ).rejects.toThrow(Error);
-    });
 
     it("should create a trade from the default address", async () => {
       const trade = Promise.resolve(tradeObject);